修改默认值

+++ qrcodecore/src/main/java/cn/bingoogolapple/qrcode/core/DisplayUtils.java
@@ -19,8 +19,8 @@ public class DisplayUtils {
         return screenResolution;
     }
 
-    public static float dp2px(Context context, float dpValue) {
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dpValue, context.getResources().getDisplayMetrics());
+    public static int dp2px(Context context, float dpValue) {
+        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dpValue, context.getResources().getDisplayMetrics());
     }
 
 }
